More JS assignments, with all extra functions coded out.
Not too long ago I was deleting and re-writting code within the "p" or the "button" element and testing each function independently. Jordan suggested coding out the extra parts so that they are still visible for the instructors, and it's still there for me to reffer back to for later assignments. Thanks Jordan!